feat: treat-list-metadata-as-stringified-json-instead-of-ipfs-uri
